+++ src/components/Navbar/Index.jsx
@@ -6,7 +6,7 @@ import { useNavigate } from "react-router-dom";
 import "./style.scss";
 
 const Navbar = (props) => {
-  const { user, userTitle } = props;
+  const { user, userTitle, children } = props;
 
   const { id, roles } = user;
   const route = (roles || []).includes("ROLE_ENGENHEIRO_CHEFE")
@@ -33,19 +33,73 @@ const Navbar = (props) => {
     <>
       <nav className={navbarClass}>
         <div className="container-fluid">
-          <a className="navbar-brand" href={route}>
+          <button
+            className="navbar-toggler"
+            type="button"
+            data-bs-toggle="collapse"
+            data-bs-target="#navbarTogglerDemo03"
+            aria-controls="navbarTogglerDemo03"
+            aria-expanded="false"
+            aria-label="Toggle navigation"
+          >
+            <span className="navbar-toggler-icon"></span>
+          </button>
+          <a className="navbar-brand ms-2" href={route}>
             <img src={logoSrc} alt="9tech-logo" width="90"></img>
           </a>
-          <div className="d-flex ms-auto me-2">
-            <span className="fs-5 fw-bold">{userTitle}</span>
+          <div className="d-flex ms-auto me-4">
+            <span className=" fs-5 fw-bold">{userTitle}</span>
           </div>
+          <div className="collapse navbar-collapse" id="navbarTogglerDemo03">
+            <ul className="navbar-nav me-auto mb-2 mb-lg-0">
+              <li className="nav-item">
+                <a
+                  className="nav-link  fw-bold"
+                  aria-current="page"
+                  href={route}
+                >
+                  Home
+                </a>
+              </li>
+              {children}
 
+              <li className="btn-group dropend">
+                <a
+                  className="nav-link dropdown-toggle fw-bold"
+                  href="#"
+                  role="button"
+                  data-bs-toggle="dropdown"
+                  aria-expanded="false"
+                >
+                  Usuários
+                </a>
+                <ul className="dropdown-menu dropdown-menu-dark">
+                  <li>
+                    <a className="dropdown-item" href="#">
+                      Cadastrar
+                    </a>
+                  </li>
+                  <li>
+                    <a className="dropdown-item" href="#">
+                      Listar
+                    </a>
+                  </li>
+                  <li>
+                    <a className="dropdown-item" href="#">
+                      Excluir
+                    </a>
+                  </li>
+                </ul>
+              </li>
+            </ul>
+          </div>
           <button
-            className="nav-link d-flex align-items-center"
+            type="button"
+            class="btn btn-outline-dark align-items-center fw-bold"
             onClick={logout}
           >
-            <h5 className="fs-5 fw-bold">SAIR</h5>
-            <i className="fa fa-right-from-bracket ms-2 fa-2x" />
+            SAIR
+            <i className="fa fa-right-from-bracket ms-2" />
           </button>
         </div>
       </nav>
